==> CODE/31.fish.rmd <==
---
title: "allfish_conditional"
author: "Sarah Roberts"
date: "2/2/2022"
output: html_document
---

First, predict out of sample by fitting the model with a training set and then predicting (conditional and traditional) the test set.

Then predict in sample by fitting the model with the entire dataset and predicting (conditional and traditional) the data used to fit the model.

Compare conditional and traditional prediction for both in and out of sample, and plot.

#packages

```{r}
library(gjam)
library(tidyverse)
library(ggthemes)
library(ggpubr)
library(ggplot2)
library(pROC)
library(gridExtra)
library(sf)
```

#functions

```{r}
r2_general <-function(preds,actual){ 
  return(1- sum((preds - actual) ^ 2)/sum((actual - mean(actual))^2))
}

RMSE_func <- function(preds, actual){
  return(sqrt(mean((actual - preds)^2)))
}

'%!in%' <- function(x,y)!('%in%'(x,y))

compareEachSpecies <- function(out, nsim = 200) {
  compare <- c()
  ydata <- out$inputs$y
  
  for (s in 1:ncol(out$inputs$y)){
    
    true <- ydata[,s]
    trad <- out$prediction$ypredMu[,s]
    
    
    # conditional prediction
    
    conditionOn <- colnames(ydata)[-s]
    conditionOnData <- select(as.data.frame(ydata), all_of(conditionOn))
    
    newdata <- list(ydataCond = conditionOnData,  nsim=nsim)
    cond_pred <- gjamPredict(output = out, newdata = newdata)
    cond <- cond_pred$sdList$yMu[,s]
    
    
    # get spe
    coni <- (ydata[,s] - cond)^2
    unci <- (ydata[,s] - trad)^2
    
    # how many spe are greater for traditional prediction
    umc  <- unci - coni
    uno  <- length(which(umc > 0))/length(umc)  # fraction larger error in uncon
    
    # get rmspe
    rmspecond <- sqrt(mean(coni))
    rmspetrad <- sqrt(mean(unci))
    
    
    # add to compare dataframe
    comp <- data.frame(species = colnames(ydata)[s],
                       "Unc RMSPE" = rmspetrad, 
                       "Con RMSPE" = rmspecond, 
                       "Perc Diff" = (rmspetrad-rmspecond)/rmspetrad * 100, 
                       "Frac u > c" = uno)
    colnames(comp) <- c("species", "Unc RMSPE", "Con RMSPE", "Perc Diff", "Frac u > c")
    
    compare <- bind_rows(compare, comp)
  }
  
  return(compare)
  
} 


```

Set up some basic parameters

```{r}
ng <- 20000
burnin <- ng/2

nsim <- 1000
```

# 1 out of sample

train on 70% and test on 30%

## 1.1 Fit model

After you fit the model, save it (and the train and test data) so you don't have to run it again.

```{r}

# load data
load("../DATA/fish/ydata.rdata")
load("../DATA/fish/xdata.rdata")


# get sqrt y
ydata <- sqrt(ydata)


# keep only data from Gulf of Maine
keep <- grep("GulfofMaine", xdata$SUBREGION2)
xdata <- xdata[keep,]
ydata <- ydata[keep,]

# keep only species with enough data
ydata <- ydata[,colSums(ydata != 0) > 150]

ynames <- colnames(ydata)
xnames <- colnames(xdata)

# # combine x and y, keep only observations from Gulf of Maine
total <- cbind(ydata, xdata)

# take out 70% for training set
smp_size <- floor(0.70 * nrow(total))
set.seed(123)
train_ind <- sample(seq_len(nrow(total)), size = smp_size)
train <- total[train_ind, ]
test <- total[-train_ind, ]

ytrain <- train[,colnames(train) %in% ynames]
xtrain <- train[,colnames(train) %in% xnames]
ytest <- test [,colnames(test ) %in% ynames]
xtest <- test [,colnames(test) %in% xnames]

#run gjam

ml <- list(ng = ng, burnin = burnin, typeNames = 'CA')
out_oos <- gjam(~BT + depth + BSAL + SST + chla + SEDSIZE + month, xdata = xtrain, ydata = ytrain, modelList = ml) 

save(out_oos, ytrain, ytest, xtrain, xtest, file='../OUT/fish-out_train_CPUE_fall.rdata')


```


## 1.2 Compare traditional and conditional predictions for out-of-sample

```{r}
load("../OUT/fish-out_train_CPUE_fall.rdata")

compare_oos <- compareEachSpecies(out_oos, nsim = nsim)


```


# 2 in sample

## 2.1 Fit model

Combine test and train data and fit model.

```{r}

# load data
load("../DATA/fish/ydata.rdata")
load("../DATA/fish/xdata.rdata")


# get sqrt y
ydata <- sqrt(ydata)


# keep only data from Gulf of Maine
keep <- grep("GulfofMaine", xdata$SUBREGION2)
xdata <- xdata[keep,]
ydata <- ydata[keep,]

# keep only species with enough data
ydata <- ydata[,colSums(ydata != 0) > 150]

ynames <- colnames(ydata)
xnames <- colnames(xdata)

# # combine x and y, keep only observations from Gulf of Maine
total <- cbind(ydata, xdata)


#run gjam
ml <- list(ng = ng, burnin = burnin, typeNames = 'CA')
out_is <- gjam(~BT + depth + BSAL + SST + chla + SEDSIZE + month, xdata = xdata, ydata = ydata, modelList = ml) 

save(out_is, xdata, ydata, file='../OUT/fish-out_full_CPUE_fall.Rdata')

```

## 2.2 Compare traditional and conditional predictions for in-sample

Condition each species on all others

```{r}
load('../OUT/fish-out_full_CPUE_fall.Rdata') #in case you don't want to run the full thing

compare_is <- compareEachSpecies(out_is, nsim = nsim)

```


# sidenote: For the two species that don't gain improvement from conditioning, are they fit in the model worse than other species?

```{r}
rmspe_is <- out_is$fit$rmspeBySpec
rmspe_oos <- out_oos$fit$rmspeBySpec

rmspe <- data.frame(species = names(rmspe_is),
                    is = rmspe_is,
                    oos = rmspe_oos)

```





#3 plot it - plot conditional and unconditional metrics together

```{r}

# format in sample metrics
all_is <- compare_is %>%
  group_by(species) %>%
  summarize(percDiffMean = mean(`Perc Diff`),
            percDiffSe = sd(`Perc Diff`)/sqrt(n()),
            percImprMean = mean(`Frac u > c`*100),
            percImprSe = sd(`Frac u > c`*100)/sqrt(n())) %>%
  pivot_longer(cols = c("percDiffMean", "percImprMean", 
                        "percDiffSe", "percImprSe")) %>%
  mutate(type = case_when(name %in% c("percDiffMean", "percImprMean") ~ "mean",
                          T ~ "se"),
         metric = case_when(name %in% c("percDiffMean", "percDiffSe") ~ "percDiff",
                            T ~ "percImpr")) %>%
  select(!name) %>%
  pivot_wider(names_from = type, values_from = value) 


# format out of sample metrics
all_oos <- compare_oos %>%
  group_by(species) %>%
  summarize(percDiffMean = mean(`Perc Diff`),
            percDiffSe = sd(`Perc Diff`)/sqrt(n()),
            percImprMean = mean(`Frac u > c`*100),
            percImprSe = sd(`Frac u > c`*100)/sqrt(n())) %>%
  pivot_longer(cols = c("percDiffMean", "percImprMean", 
                        "percDiffSe", "percImprSe")) %>%
  mutate(type = case_when(name %in% c("percDiffMean", "percImprMean") ~ "mean",
                          T ~ "se"),
         metric = case_when(name %in% c("percDiffMean", "percDiffSe") ~ "percDiff",
                            T ~ "percImpr")) %>%
  select(!name) %>%
  pivot_wider(names_from = type, values_from = value) 

# calculate number of zeros
yzeros <- ydata
yzeros[yzeros > 0] <- 1
zeros <- as.data.frame(colMeans(yzeros))
zeros$species <- rownames(zeros)
colnames(zeros) <- c("abun", "species")

# join everything together
full <- full_join(all_is, all_oos, by = c("species", "metric")) %>%
  full_join(zeros, "species") %>%
  pivot_longer(cols = c("mean.x", "mean.y")) %>%
  mutate(Species = case_when(species == "ACADIAN.REDFISH" ~ "Acadian Redfish",
                             species == "AMERICAN.LOBSTER" ~ "American Lobster",
                             species == "AMERICAN.PLAICE" ~ "American Plaice",
                             species == "ATLANTIC.COD" ~ "Atlantic Cod",
                             species == "ATLANTIC.HERRING" ~ "Atlantic Herring",
                             species == "BARNDOOR.SKATE" ~ "Barndoor Skate",
                             species == "BUTTERFISH" ~ "Butterfish",
                             species == "FOURSPOT.FLOUNDER" ~ "Four-spot Flounder",
                             species == "HADDOCK" ~ "Haddock",
                             species == "LITTLE.SKATE" ~ "Little Skate",
                             species == "LONGFIN.SQUID" ~ "Longfin Squid",
                             species == "LONGHORN.SCULPIN" ~ "Longhorn Sculpin",
                             species == "NORTHERN.SEAROBIN" ~ "Northern Searobin",
                             species == "NORTHERN.SHORTFIN.SQUID" ~ "Northern Shortfin Squid",
                             species == "POLLOCK" ~ "Pollock",
                             species == "RED.HAKE" ~ "Red Hake",
                             species == "SEA.SCALLOP" ~ "Sea Scallop",
                             species == "SILVER.HAKE" ~ "Silver Hake",
                             species == "SPINY.DOGFISH" ~ "Spinny Dogfish",
                             species == "SUMMER.FLOUNDER" ~ "Summer Flounder",
                             species == "WHITE.HAKE" ~ "White Hake",
                             species == "WINTER.FLOUNDER" ~ "Winter Flounder",
                             species == "WINTER.SKATE" ~ "Winter Skate",
                             species == "YELLOWTAIL.FLOUNDER" ~ "Yellowtail Flounder"),
         group = case_when(name == "mean.x" ~ "In-sample",
                           name == "mean.y" ~ "Out-of-sample")) 


b <- ggplot(filter(full, metric == "percDiff")) + 
  # geom_hline(yintercept = 50, color = "deepskyblue", alpha = 0.5) +
  geom_hline(yintercept = 0, color = "black", alpha = 0.5) +
  geom_point(aes(x = reorder(Species, abun), y = value, color = metric, shape = name, size = name)) +
  scale_color_manual(#labels = c("% Improvement in RMSPE \nfrom Conditioning",
    #          "Fraction of Predictions\nImproved by Conditioning"),
    values = c("coral3","deepskyblue"), guide = "none") +
  scale_shape_manual(values = c(16, 1), guide = "none") +
  scale_size_manual(values = c(2.2, 3), guide = "none") +
  theme_bw() +
  theme(strip.background = element_blank(),
        panel.grid.minor = element_blank(),
        panel.spacing = unit(1.5, "lines"),
        plot.margin = unit(c(1,.1,0,0), "cm"),
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1,
                                   size = 9),
        axis.text.y = element_blank(),
        axis.title = element_text(size = 9)) +
  coord_cartesian(ylim = c(0, 25), clip = "on") +
  labs(x = "", y = "")




a <- ggplot() + 
  # geom_hline(yintercept = 50, color = "deepskyblue", alpha = 0.5) +
  geom_hline(yintercept = 0, color = "black", alpha = 0.5) +
  geom_boxplot(data = filter(full, metric == "percDiff"),
               aes(x = group, y = value, alpha = name),
               color = "coral3",
               fill = "coral3") +
  # geom_boxplot(data = filter(full, metric == "percImpr"),
  #              aes(x = group, y = value, alpha = name),
  #              fill = "deepskyblue",
  #              color = "deepskyblue") +
  scale_alpha_manual(values = c(.6, 0)) +
  coord_cartesian(ylim = c(0, 25)) +
  theme_bw() +
  theme(strip.background = element_blank(),
        panel.grid.minor = element_blank(),
        panel.spacing = unit(1.5, "lines"),
        plot.margin = unit(c(0, 0, 1, 0), "cm"),
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1,
                                   size = 9),
        axis.text.y = element_text(size = 9),
        axis.title = element_text(size = 9),
        legend.position = "none") +
  labs(x = "", y = "Percent improvement \n in RMSPE")
  



ggpubr::ggarrange(a, b, nrow = 1,
                  labels = "auto",
                  widths = c(0.3, 1),
                  label.x = c(0.13, 0), label.y = .95,
                  font.label = list(size = 10),
                  align = "h")


ggsave(filename = "../OUT/figures/metrics.png",
       height = 110, width = 180, units = "mm", dpi = 600)

```

#4 Map

```{r}
## Map ----

# map in sample
output <- out_is
ydata <- output$inputs$y
xdata <- output$inputs$xdata



# format xdata with lat and longs
total1 <- read.csv("../DATA/fish/total_with_env.csv") #this has the missing lat longs in it 

colnames(total1) <- gsub("\\_", ".", colnames(total1)) #gjam doesnt like _ in column names 
total1 <- total1[total1$year>1986,] 
total1 <- total1[total1$SEASON == "FALL",]

total1$SEDSIZE <- ifelse(total1$SEDIMENT ==  '0 - 0.03 Silt/Mud', .03, 
                        ifelse(total1$SEDIMENT ==  '0.03 - 0.17 Sand', .17,
                               ifelse(total1$SEDIMENT ==  '0.17 - 0.35 Sand',.35, 
                                      ifelse(total1$SEDIMENT ==  '0.35 - 0.36 Sand', .36,
                                             ifelse(total1$SEDIMENT ==  '0.36 - 0.48 Sand', .48, ifelse(total1$SEDIMENT ==  '0.48+ Coarse Sand to Gravel',.50,NA))))))

xnames <- c("depth", "BT", "BSAL", "SSAL", "SST", "chla", "month", "season", "SEDSIZE", "SUBREGION2", "haulid", "lat", "lon", "stratum", "stratumarea", "year")
xdata1 <- total1[colnames(total1) %in% xnames]
xdata1 <- xdata1[xdata1$SUBREGION2 == "Gulf of Maine",]
xdata1$SUBREGION2 <- gsub("Gulf of Maine", "GulfofMaine", xdata1$SUBREGION2 )

xdata1$season <- as.factor(xdata1$season)
xdata1$SUBREGION2 <- as.factor(xdata1$SUBREGION2)

xdata <- left_join(xdata, xdata1,by = c("depth", "BT", "BSAL", "SSAL", "SST", "chla", "month", "season", "SEDSIZE", "SUBREGION2"))



### get predictions together ----

## observed
total_strat_obs <- cbind(xdata, ydata) 
total_strat_obs <- total_strat_obs %>% 
  pivot_longer(cols = colnames(ydata), names_to = "spp", values_to = "biomass") %>%
  filter(spp == "ATLANTIC.COD")

total_strat_obs$pred_typ <- "observed"

## conditional prediction
conditionOn <- ydata[,-grep("ATLANTIC.COD", colnames(ydata))]

newdata <- list(ydataCond = conditionOn,  nsim=2000)
cond_pred <- gjamPredict(output = out_is, newdata = newdata)
pred_con <- cond_pred$sdList$yMu[,grep("ATLANTIC.COD", colnames(cond_pred$sdList$yMu))]
      
total_strat_con <- cbind(xdata, pred_con) 
total_strat_con <- total_strat_con %>%
  mutate(spp = "ATLANTIC.COD",
         biomass = pred_con) %>%
  select(!pred_con)


total_strat_con$pred_typ <- "conditional"

## traditional prediction
pred_uncon <- out_is$prediction$ypredMu
total_strat_uncon <- cbind(xdata, pred_uncon) 
total_strat_uncon <- total_strat_uncon %>% 
  pivot_longer(cols = colnames(pred_uncon), names_to = "spp", values_to = "biomass") %>%
  filter(spp == "ATLANTIC.COD")


total_strat_uncon$pred_typ <- "unconditional"

total_strat <- rbind(total_strat_obs, total_strat_con, total_strat_uncon)


## combine
total_strat <- total_strat %>% 
  group_by(season, spp, year, stratum, stratumarea, pred_typ) %>% 
  summarise(biomass = mean(biomass)) %>% 
  mutate(biomass_area = biomass*stratumarea,
         pred_type = factor(pred_typ, levels = c("observed", "unconditional", "conditional")))


### plot ----
strat <- sf::st_read("../DATA/fish/stratum/BTS_Strata.shp")
#plot(strat)

strat <- left_join(strat, total_strat, by = c("STRATA" = "stratum"))
st_crs(strat) = 4326 #this is WGS 1984


### plot ----
library(rnaturalearth)
world <- ne_countries(scale = "medium", returnclass = "sf")


#### difference maps 
strat_sub <- strat %>% 
  dplyr::select(-pred_typ, -biomass) %>% 
  pivot_wider(names_from = "pred_type", values_from = "biomass_area") %>% 
  mutate(sub_con = conditional - observed, 
         sub_uncon = unconditional - observed) %>%
  pivot_longer(cols = c("sub_uncon", "sub_con", "observed"))

strat_sub <- st_as_sf(strat_sub)
strat_sub <- strat_sub %>% filter(name %in% c("sub_uncon", "sub_con", "observed"))
st_crs(strat_sub) = 4326
strat_sub$name1 <- as.factor(strat_sub$name)


# just for the maps, set limits at 2000 and -1000
# strat_sub$value[which(strat_sub$value > 2000)] <- 2000
# strat_sub$value[which(strat_sub$value < -1000)] <- -1000



base1 <- ggplot() + 
  geom_sf(data = world) + 
  labs(fill = "Atlantic cod \nbiomass (kg/km2)")+
  theme_bw() +
  theme(panel.grid = element_blank(),
        strip.background = element_blank(),
        strip.text = element_text(hjust = 0),
        plot.title = element_text(size = 9,
                             hjust = 0.5),
        axis.title = element_blank(),
        axis.text = element_text(size = 9),
        plot.margin = margin(5, 1, 3, 1),
        legend.key.height = unit(.5, "cm"),
        legend.key.width = unit(.5, "cm"),
        legend.title = element_text(size = 9,
                                    hjust = 0.5),
        legend.direction = "vertical",
        legend.text= element_text(size = 9)) +
  scale_fill_gradientn(colors = c("white", "yellow", "firebrick2"), 
                      #midpoint = 0, breaks = c(-1000,0,1000,2000),
                      #limits = c(-1000, 2000),
                      guide = guide_colorbar(title.position = "top")) +
  scale_y_continuous(breaks = c(40, 43, 46)) +
  scale_x_continuous(breaks = c(-66, -70, -74))

base2 <- ggplot() + 
  geom_sf(data = world) + 
  labs(fill = "Difference in \nbiomass (kg/km2)")+
  theme_bw() +
  theme(panel.grid = element_blank(),
        strip.background = element_blank(),
        strip.text = element_text(hjust = 0),
        plot.title = element_text(size = 9,
                             hjust = 0.5),
        axis.title = element_blank(),
        axis.text = element_text(size = 9),
        plot.margin = margin(0, 1, 3, 1),
        legend.key.height = unit(.5, "cm"),
        legend.key.width = unit(.5, "cm"),
        legend.title = element_text(size = 9,
                                    hjust = 0.5,
                                    vjust = 1),
        legend.direction = "vertical",
        legend.text = element_text(size = 9)) +
  scale_fill_gradient2(low = "darkblue", high = "red4", mid = "white", 
                      midpoint = 0, breaks = c(-1000,0,1000,2000),
                      limits = c(-1050, 2000),
                      guide = guide_colorbar(title.position = "top")) +
  scale_y_continuous(breaks = c(40, 43, 46),
                     labels = c("     ", "     ", "     ")) +
  scale_x_continuous(breaks = c(-66, -70, -74),
                     labels = c("", "", ""))


a <- base1 +
  geom_sf(data = filter(strat_sub, 
                        strat_sub$spp == "ATLANTIC.COD",
                        strat_sub$year == 2012,
                        name == "observed"),
          aes(group = as.factor(value), fill = value)) +
  coord_sf(xlim=c(-74, -65), ylim=c(40,46), expand = TRUE) +
  labs(title = "Observed")
  
b <- base2 +
  geom_sf(data = filter(strat_sub, 
                        strat_sub$spp == "ATLANTIC.COD",
                        strat_sub$year == 2012,
                        name == "sub_uncon"),
          aes(group = as.factor(value), fill = value)) +
  coord_sf(xlim=c(-74, -65), ylim=c(40,46), expand = TRUE) +
  labs(title = "Error from traditional prediction") #+
  #theme(axis.text = element_blank())
  
c <- base2 +
  geom_sf(data = filter(strat_sub, 
                        strat_sub$spp == "ATLANTIC.COD",
                        strat_sub$year == 2012,
                        name == "sub_con"),
          aes(group = as.factor(value), fill = value)) +
  coord_sf(xlim=c(-74, -65), ylim=c(40,46), expand = TRUE) +
  labs(title = "Error from conditional prediction") #+
  #theme(axis.text = element_blank())

# add panel comparing b and c? No.
# tmp <- strat_sub %>%
#   filter(name %in% c("sub_con", "sub_uncon"), 
#                         strat_sub$spp == "ATLANTIC.COD",
#                         strat_sub$year == 2012) %>%
#   select(!c("conditional", "unconditional", "name")) %>%
#   pivot_wider(names_from = "name1", values_from = "value") %>%
#   mutate(diff = sub_uncon - sub_con)
# 
# d <- base2 +
#   geom_sf(data = tmp,
#           aes(fill = diff)) +
#   coord_sf(xlim=c(-74, -65), ylim=c(40,46), expand = TRUE) +
#   labs(title = "Traditional prediction error compared to \nconditional prediction error") #+
#   #theme(axis.text = element_blank())


### I can't get panels lined up if they have separate legends, so I'm
#   arranging this manually :(

# this makes the figure with panels aligned
ggpubr::ggarrange(a, b, c, ncol = 1,
                  labels = "auto", 
                  label.y = 1, label.x = 0.15,
                  legend = "none",
                  font.label = list(size = 10),
                  align = "hv")

ggsave("../OUT/figures/codmaps.png", width = 80, height=180, units=c("mm"), 
       dpi = 600)

# this makes the legends
top <- ggpubr::ggarrange(a, labels = "a", legend = "right")
bottom <- ggpubr::ggarrange(b, c, labels = c("b", "c"), 
                            legend = "right", ncol = 1,
                            common.legend = T)

ggpubr::ggarrange(top, bottom, ncol = 1, heights = c(1, 2))

ggsave("../OUT/figures/codmaps-legends.png", width = 80, height=180, units=c("mm"), 
       dpi = 600)

```
